Add --bfpath option
Currently, `--bf` requires access to the managed repository.
If, however, a copy of the necessary files are already available
locally, it would be necessary to first determine where they live
on the server, and place the files in exactly that location.

The `--bfpath` option takes the absolute or relative path to the
files and uses that rather than inspect the OMERO repository.